Complete Spring Boot Projects/rentvideo/src/main/java/com/crio/rentvideo/controllers/VideoController.java
package com.crio.rentvideo.controllers;

import com.crio.rentvideo.models.Video;
import com.crio.rentvideo.services.VideoService;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/videos")
public class VideoController {

    @Autowired
    private VideoService videoService;

    @GetMapping
    public List<Video> getAllVideos() {
        return videoService.getAllVideos();
    }

    @PostMapping("/manage")
    public ResponseEntity<Video> addVideo(@Valid @RequestBody Video video) {
        return new ResponseEntity<>(videoService.addVideo(video), HttpStatus.CREATED);
    }

    @PutMapping("/manage/{id}")
    public ResponseEntity<Video> updateVideo(@PathVariable Long id, @Valid @RequestBody Video video) {
        return ResponseEntity.ok(videoService.updateVideo(id, video));
    }

    @DeleteMapping("/manage/{id}")
    public ResponseEntity<String> deleteVideo(@PathVariable Long id) {
        videoService.deleteVideo(id);
        return ResponseEntity.ok("Video deleted successfully");
    }
}
